# myexample/fastapi-whiskey-app/app/models/distillery.py
from sqlalchemy import Column, Integer, String, ForeignKey
from sqlalchemy.orm import relationship
from app.database import Base

class Distillery(Base):
    __tablename__ = 'distilleries'

    id = Column(Integer, primary_key=True, index=True)
    name = Column(String, nullable=False)
    location = Column(String, nullable=False)

    whiskies = relationship("Whisky", back_populates="distillery")